requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.6.5
appnope==0.1.2
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
decorator==5.1.0
deprecation==2.1.0
finnhub-python==2.4.5
flask-cors==3.0.8
flask-login==0.5.0
flask-migrate==3.0.1
flask-sqlalchemy==2.5.1
flask-wtf==0.15.1
flask==2.0.1
frozendict==2.0.7
greenlet==1.1.0
gunicorn==20.1.0
idna==3.3
ipython==7.29.0
itsdangerous==2.0.1
jedi==0.18.0
jinja2==3.0.1
mako==1.1.4
markupsafe==2.0.1
matplotlib-inline==0.1.3
numpy==1.21.4
packaging==21.2; python_version >= '3.6'
pandas==1.3.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.4.0
prompt-toolkit==3.0.22
ptyprocess==0.7.0
pyex==0.5.0
pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz-deprecation-shim==0.1.0.post0
pytz==2021.3
requests==2.26.0
six==1.15.0
socketio-client-nexus==0.7.6
sqlalchemy==1.4.19
sseclient==0.0.27
temporal-cache==0.1.4
traitlets==5.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
wcwidth==0.2.5
websocket-client==1.2.1
werkzeug==2.0.1
wtforms==2.3.3
